refactor(parser): modularize pipeline and add parser usage guide
