+++ parser/src/io/decoder/errors.ts
@@ -1,0 +1,6 @@
+export class DecoderError extends Error {
+  constructor(message: string, options?: ErrorOptions) {
+    super(message, options);
+    this.name = 'DecoderError';
+  }
+}
