+++ parser/src/plugins/index.ts
@@ -1,4 +1,14 @@
-export { PluginRegistry } from './registry.js';
+import { PluginRegistry } from './registry.js';
+import type { ParserPlugin } from './types.js';
+
+export { PluginRegistry };
+
+export function createPluginRegistry(
+  plugins: readonly ParserPlugin[] = []
+): PluginRegistry {
+  return new PluginRegistry(plugins);
+}
+
 export type { ExtensionCollector } from './registry.js';
 export type {
   ExtensionEvaluation,
